Fix minor issues in ProductionHistory

+++ src/main/java/org/chainoptim/features/scanalysis/production/productionhistory/model/ProductionHistory.java
@@ -13,6 +13,6 @@ import java.util.Map;
 public class ProductionHistory {
 
     private LocalDateTime startDate;
-    private Map<Integer, DailyProductionRecord> dailyProductionRecords;
+    private Map<Float, DailyProductionRecord> dailyProductionRecords;
 
 }
